Rename files for reflection probe
