+++ projects/PathosEngine/PathosEngine.vcxproj.filters
@@ -372,10 +372,10 @@
     <ClCompile Include="src\pathos\scene\sky_ansel_component.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="src\pathos\scene\light_probe_component.cpp">
+    <ClCompile Include="src\pathos\scene\reflection_probe_component.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="src\pathos\scene\light_probe_actor.cpp">
+    <ClCompile Include="src\pathos\scene\reflection_probe_actor.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="src\pathos\render\visualize_light_probe.cpp">
@@ -830,10 +830,10 @@
     <ClInclude Include="src\badger\math\vector_math.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="src\pathos\scene\light_probe_component.h">
+    <ClInclude Include="src\pathos\scene\reflection_probe_component.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="src\pathos\scene\light_probe_actor.h">
+    <ClInclude Include="src\pathos\scene\reflection_probe_actor.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="src\pathos\render\visualize_light_probe.h">
