+++ projects/PathosEngine/src/pathos/render/indirect_lighting.cpp
@@ -9,7 +9,7 @@
 #include "pathos/render/scene_proxy.h"
 #include "pathos/render/render_target.h"
 #include "pathos/scene/camera.h"
-#include "pathos/scene/light_probe_component.h"
+#include "pathos/scene/reflection_probe_component.h"
 #include "pathos/scene/irradiance_volume_actor.h"
 #include "pathos/mesh/geometry.h"
 #include "pathos/mesh/geometry_primitive.h"
