+++ projects/PathosEngine/src/pathos/scene/reflection_probe_component.h
